+++ com.unity.render-pipelines.high-definition/Runtime/Lighting/LightLoop/LightLoop.cs
@@ -3609,8 +3609,7 @@ namespace UnityEngine.Rendering.HighDefinition
 
             // Misc
             cb._EnableSSRefraction = hdCamera.frameSettings.IsEnabled(FrameSettingsField.Refraction) ? 1u : 0u;
-            cb._CloudShadowOpacity = hdCamera.lightingSky.cloudLayer == null ? 1.0f : 1.0f - hdCamera.lightingSky.cloudLayer.shadowsOpacity.value;
-            cb._CloudShadowScale = hdCamera.lightingSky.cloudLayer == null ? 1.0f : hdCamera.lightingSky.cloudLayer.shadowsScale.value;
+            cb._CloudShadowScale = hdCamera.lightingSky.cloudLayer == null ? 1.0f : hdCamera.lightingSky.cloudLayer.shadowsTiling.value;
         }
 
         void PushLightDataGlobalParams(CommandBuffer cmd)